Fix g++-13 useless-cast on the priced-seed count
CI's g++-13 rejects static_cast<std::uint64_t>(vector::size()) under
-Werror=useless-cast: on LP64 Linux size_t and uint64_t are the same type, so
the cast is a no-op. g++-16 and clang keep quiet because on macOS the two types
differ, which is why the local GCC build did not catch it. Assign the size
directly, matching how MultiSeedSummary::seed_count is set.

+++ src/experiments/heston_simulation_experiments.cpp
@@ -104,7 +104,7 @@ Result<SchemeCell> run_cell(const MarketState& market,
 
     cell.negative_fraction = negative_fraction_sum / static_cast<double>(seeds.size());
     cell.mean_runtime_seconds = runtime_sum / static_cast<double>(seeds.size());
-    cell.priced_seeds = static_cast<std::uint64_t>(replications.size());
+    cell.priced_seeds = replications.size();
 
     if (replications.size() >= 2) {
         auto summary = summarize_seeds(replications, reference);
